requirements.txt
pandas
jinja2
aiogram==3.13.1
python-dotenv
Pillow
PyYAML
html2image
